requirements.txt
absl-py==0.11.0
actionlib==1.13.2
aiohttp==3.7.3
aiohttp-cors==0.7.0
aioredis==1.3.1
alabaster==0.7.8
angles==1.9.13
apturl==0.5.2
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
atari-py==0.2.6
attrs==19.3.0
autobahn==17.10.1
Automat==0.8.0
Babel==2.6.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.9.3
bleach==3.2.1
blessings==1.7
blinker==1.4
bondpy==1.8.6
breezy==3.0.2
Brlapi==0.7.0
cachetools==4.2.0
camera-calibration==1.15.2
camera-calibration-parsers==1.12.0
catkin==0.8.9
catkin-pkg==0.4.23
catkin-pkg-modules==0.4.23
catkin-tools==0.5.0
cbor==1.0.0
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
Click==7.0
cloudpickle==1.6.0
colorama==0.4.3
colorful==0.5.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
controller-manager==0.19.4
controller-manager-msgs==0.19.4
cryptography==2.8
cupshelpers==1.0
cv-bridge==1.15.0
cycler==0.10.0
Cython==0.29.14
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
Deprecated==1.2.7
diagnostic-analysis==1.10.3
diagnostic-common-diagnostics==1.10.3
diagnostic-updater==1.10.3
distro==1.4.0
distro-info===0.23ubuntu1
dm-tree==0.1.5
docopt==0.6.2
docutils==0.16
dulwich==0.19.15
duplicity==0.8.12.0
dynamic-reconfigure==1.7.1
empy==3.3.2
entrypoints==0.3
fasteners==0.14.1
fastimport==0.9.8
filelock==3.0.12
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gazebo-plugins==2.9.1
gazebo-ros==2.9.1
gencpp==0.6.5
geneus==3.0.0
genlisp==0.4.18
genmsg==0.5.16
gennodejs==2.0.2
genpy==0.6.14
gin-config==0.1.3
google==3.0.0
google-api-core==1.24.1
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
googleapis-common-protos==1.52.0
gpg===1.13.1-unknown
gpustat==0.6.0
grpcio==1.32.0
gym==0.18.0
h5py==2.10.0
hiredis==1.1.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
image-geometry==1.15.0
imagesize==1.2.0
incremental==16.10.1
interactive-markers==1.12.0
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.2
jedi==0.18.0
Jinja2==2.11.2
joblib==1.0.0
joint-state-publisher==1.15.0
joint-state-publisher-gui==1.15.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
laser-geometry==1.6.5
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.6.2
lz4==3.0.2+dfsg
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.3
MarkupSafe==1.1.0
matplotlib==3.1.2
message-filters==1.15.9
mistune==0.8.4
monotonic==1.5
mpi4py==3.0.3
msgpack==1.0.2
multidict==5.1.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
netifaces==0.10.4
nose==1.3.7
notebook==6.1.6
numpy==1.19.4
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
olefile==0.46
opencensus==0.7.11
opencensus-context==0.1.2
opencv-python==4.5.1.48
opencv-python-headless==4.3.0.36
opt-einsum==3.3.0
osrf-pycommon==0.1.9
packaging==20.8
pandas==1.2.0
pandocfilters==1.4.3
paramiko==2.6.0
parso==0.8.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.14.0
psutil==5.5.1
ptyprocess==0.6.0
py-spy==0.3.3
py-ubjson==0.14.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycparser==2.20
pycryptodomex==3.6.1
pycups==1.9.73
pydot==1.4.1
PyGithub==1.43.7
pyglet==1.5.0
Pygments==2.7.3
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
PyOpenGL==3.1.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pypng==0.0.20
PyQRCode==1.2.1
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.17.3
python-apt==2.0.0+ubuntu0.20.4.3
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-gitlab==2.0.1
python-gnupg==0.4.5
python-qt-binding==0.4.3
python-snappy==0.5.3
PyTrie==0.2
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==20.0.0
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
qtconsole==5.0.1
QtPy==1.9.0
ray==1.1.0
redis==3.5.3
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
resource-retriever==1.12.6
roman==2.0.0
rosbag==1.15.9
rosboost-cfg==1.15.7
rosclean==1.15.7
roscreate==1.15.7
rosdep==0.20.0
rosdep-modules==0.20.0
rosdistro==0.8.3
rosdistro-modules==0.8.3
rosgraph==1.15.9
roslaunch==1.15.9
roslib==1.15.7
roslint==0.12.0
roslz4==1.15.9
rosmake==1.15.7
rosmaster==1.15.9
rosmsg==1.15.9
rosnode==1.15.9
rosparam==1.15.9
rospkg==1.2.9
rospkg-modules==1.2.9
rospy==1.15.9
rosservice==1.15.9
rostest==1.15.9
rostopic==1.15.9
rosunit==1.15.7
roswtf==1.15.9
rqt-action==0.4.9
rqt-bag==0.5.0
rqt-bag-plugins==0.5.0
rqt-console==0.4.11
rqt-dep==0.4.10
rqt-graph==0.4.14
rqt-gui==0.5.2
rqt-gui-py==0.5.2
rqt-image-view==0.4.16
rqt-launch==0.4.9
rqt-logger-level==0.4.11
rqt-moveit==0.5.9
rqt-msg==0.4.9
rqt-nav-view==0.5.7
rqt-plot==0.4.12
rqt-pose-view==0.5.10
rqt-publisher==0.4.9
rqt-py-common==0.5.2
rqt-py-console==0.4.9
rqt-reconfigure==0.5.3
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.13
rqt-robot-steering==0.5.12
rqt-runtime-monitor==0.5.8
rqt-rviz==0.6.1
rqt-service-caller==0.4.9
rqt-shell==0.4.10
rqt-srv==0.4.8
rqt-tf-tree==0.6.2
rqt-top==0.4.9
rqt-topic==0.4.12
rqt-web==0.4.9
rsa==4.6
rviz==1.14.4
scikit-learn==0.24.0
scipy==1.5.4
screen-resolution-extra==0.0.0
seaborn==0.11.1
SecretStorage==2.3.1
Send2Trash==1.5.0
sensor-msgs==1.13.0
service-identity==18.1.0
simplejson==3.16.0
sip==4.19.21
six==1.15.0
smach==2.5.0
smach-ros==2.5.0
smclib==1.8.6
soupsieve==2.1
Sphinx==1.8.5
stable-baselines==2.10.1
systemd-python==234
tabulate==0.8.7
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.1
tensorflow==2.4.0
tensorflow-estimator==2.4.0
tensorflow-probability==0.12.1
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
tf==1.13.2
tf-agents==0.5.0
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.5
tf2-kdl==0.7.5
tf2-py==0.7.5
tf2-ros==0.7.5
threadpoolctl==2.1.0
topic-tools==1.15.9
torch==1.7.1
tornado==6.1
traitlets==5.0.5
Twisted==18.9.0
txaio==2.10.0
typing-extensions==3.7.4.3
u-msgpack-python==2.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
vcstools==0.1.42
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
wsaccel==0.6.2
wstool==0.1.18
xacro==1.14.5
xkit==0.0.0
yarl==1.6.3
zope.interface==4.7.1
